In debug, the cubes stutter. In release, the camera stutters. What on earth?

+++ brasstacks/include/brasstacks/Meshes/VertexBuffer.hpp
@@ -11,7 +11,7 @@ class VertexBuffer {
 public:
     virtual void bind() = 0;
 
-    virtual void set_buffer(void *buffer, std::uint32_t size) = 0;
+    virtual void set_buffer(void *buffer, std::size_t size) = 0;
     virtual void set_indices(const Mesh::Face *faces,
                              const std::size_t face_count) = 0;
